fixup! feat: japanese input
