+++ src/components/Pega_Extensions_JapaneseInput/Docs.mdx
@@ -5,7 +5,14 @@ import * as DemoStories from './demo.stories';
 
 # Overview
 
-This component that converts Japanese input, such as kana characters and full-width characters.
+Japanese input component has the following features.
+
+- Lowercase to Uppercase  
+  Converts Lowercase to Uppercase, for when the client inputs lowercase and the internal character handles it as uppercase.
+- Hiragana to Katakana  
+  Japanese has two types of characters, Hiragana and Katakana, which correspond one-to-one. Hiragana input is converted to Katakana.
+- Full-width to Half-width  
+  In Japanese, characters are generally handled as multi-byte full-width characters rather than single-byte half-width characters, and for this reason Japanese keyboards have a function to input characters in full-width. Some full-width characters have the same meaning in half-width, and these are converted from full-width to half-width.
 
 <Primary />
