+++ src/components/Pega_Extensions_JapaneseInput/utils.ts
@@ -1,0 +1,40 @@
+export const convertHiraganaToKatakana = (str: string) => {
+  return str.replace(/[\u3041-\u3096]/g, match => {
+    return String.fromCharCode(match.charCodeAt(0) + 0x60);
+  });
+};
+
+export const fullWidthToHalfWidth = (str: string): string => {
+  str = str.replace(/[Ａ-Ｚａ-ｚ０-９]/g, match => {
+    return String.fromCharCode(match.charCodeAt(0) - 0xfee0);
+  });
+  // prettier-ignore
+  const kanaMap: Record<string, string> = {
+    ガ: 'ｶﾞ', ギ: 'ｷﾞ', グ: 'ｸﾞ', ゲ: 'ｹﾞ', ゴ: 'ｺﾞ',
+    ザ: 'ｻﾞ', ジ: 'ｼﾞ', ズ: 'ｽﾞ', ゼ: 'ｾﾞ', ゾ: 'ｿﾞ',
+    ダ: 'ﾀﾞ', ヂ: 'ﾁﾞ', ヅ: 'ﾂﾞ', デ: 'ﾃﾞ', ド: 'ﾄﾞ',
+    バ: 'ﾊﾞ', ビ: 'ﾋﾞ', ブ: 'ﾌﾞ', ベ: 'ﾍﾞ', ボ: 'ﾎﾞ',
+    パ: 'ﾊﾟ', ピ: 'ﾋﾟ', プ: 'ﾌﾟ', ペ: 'ﾍﾟ', ポ: 'ﾎﾟ',
+    ヴ: 'ｳﾞ', ヷ: 'ﾜﾞ', ヺ: 'ｦﾞ',
+    ア: 'ｱ', イ: 'ｲ', ウ: 'ｳ', エ: 'ｴ', オ: 'ｵ',
+    カ: 'ｶ', キ: 'ｷ', ク: 'ｸ', ケ: 'ｹ', コ: 'ｺ',
+    サ: 'ｻ', シ: 'ｼ', ス: 'ｽ', セ: 'ｾ', ソ: 'ｿ',
+    タ: 'ﾀ', チ: 'ﾁ', ツ: 'ﾂ', テ: 'ﾃ', ト: 'ﾄ',
+    ナ: 'ﾅ', ニ: 'ﾆ', ヌ: 'ﾇ', ネ: 'ﾈ', ノ: 'ﾉ',
+    ハ: 'ﾊ', ヒ: 'ﾋ', フ: 'ﾌ', ヘ: 'ﾍ', ホ: 'ﾎ',
+    マ: 'ﾏ', ミ: 'ﾐ', ム: 'ﾑ', メ: 'ﾒ', モ: 'ﾓ',
+    ヤ: 'ﾔ', ユ: 'ﾕ', ヨ: 'ﾖ',
+    ラ: 'ﾗ', リ: 'ﾘ', ル: 'ﾙ', レ: 'ﾚ', ロ: 'ﾛ',
+    ワ: 'ﾜ', ヲ: 'ｦ', ン: 'ﾝ',
+    ァ: 'ｧ', ィ: 'ｨ', ゥ: 'ｩ', ェ: 'ｪ', ォ: 'ｫ',
+    ッ: 'ｯ', ャ: 'ｬ', ュ: 'ｭ', ョ: 'ｮ',
+    '。': '｡', '、': '､', ー: 'ｰ', '「': '｢', '」': '｣', '・': '･'
+  };
+  const reg = new RegExp(`(${Object.keys(kanaMap).join('|')})`, 'g');
+  return str
+    .replace(reg, match => {
+      return kanaMap[match];
+    })
+    .replace(/゛/g, 'ﾞ')
+    .replace(/゜/g, 'ﾟ');
+};
